requirements.txt
fastapi
uvicorn
python-jose
motor
python-dotenv
python-multipart